pyproject.toml
[project]
name = "dailytech"
version = "0.1.0"
description = "Echo-Tech Daily — Automated AI news briefing pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "feedparser>=6.0.11",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "ty>=0.0.1a6",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = []

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
